Refactor: SearchResultView 기능에 따른 컴포넌트 분리 - SearchResultContainer 생성 및 데이터 호출 기능 이전 #104
